se modificaron los metodos de servicio y controladores para que identifique si el tenant o _id sean correctos antes de realizar una accion

+++ src/services/egresos.service.js
@@ -4,6 +4,14 @@ const terceroModel = require("../models/terceros.Model");
 const counterService = require("../services/counter.service");
 
 const obtenerEgresos = async (tenantId) => {
+
+  try {
+    // Verificar que el tenantId coincide con el tenantId de los egresos
+  const egresosExisten = await Egreso.exists({ tenantId });
+
+  if (!egresosExisten) {
+    throw new Error("TenantId proporcionado no es valido o no se encuentra en la base de datos");
+  }
   const egresos = await Egreso.find({ tenantId })
     .populate({
       path: "categoria",
@@ -14,9 +22,22 @@ const obtenerEgresos = async (tenantId) => {
       model: terceroModel,
     });
   return egresos;
+  } catch (error) {
+    throw error; // Propaga el error para que sea manejado en el controlador
+  }
+  
 };
 
 const obtenerEgresoPorId = async (egresoId, tenantId) => {
+
+  try {
+
+    // Verificar que el tenantId coincide con el tenantId del egreso
+  const egresoExistente = await Egreso.findOne({ _id: egresoId, tenantId });
+
+  if (!egresoExistente) {
+    throw new Error("TenantId proporcionado no es valido o no se encuentra en la base de datos");
+  }
   const egreso = await Egreso.findById({ _id: egresoId, tenantId })
     .populate({
       path: "categoria",
@@ -27,12 +48,27 @@ const obtenerEgresoPorId = async (egresoId, tenantId) => {
       model: terceroModel,
     });
   return egreso;
+  } catch (error) {
+    if (error.name === 'CastError' && error.path === '_id') {
+      throw new Error("_id proporcionado no es válido o no se encontro en la base de datos");
+    } else {
+      throw error; // Propaga el error para que sea manejado en el controlador
+    }
+  }
+
+  
 };
 
 const guardarEgreso = async (egreso, tenantId) => {
   // Asignar el egresoId y el tenantId al egreso
   egreso.tenantId = tenantId;
   egreso.egresoId = 0;
+
+  // Validar que el objeto egreso tenga la estructura correcta y campos requeridos
+  if (!egreso || !egreso.detalle || !egreso.valor) {
+    throw new Error("El objeto egreso no es valido o no contiene campos requeridos");
+  }
+
   // Crear nuevo egreso
   const nuevoEgreso = new Egreso(egreso);
 
@@ -52,29 +88,55 @@ const actualizarEgresoId = async (tenantId, idEgreseActual) => {
 };
 
 const eliminarEgresoPorId = async (egresoId, tenantId) => {
-  const egreso = await Egreso.findOne({ _id: egresoId, tenantId });
+  try {
+     // Verificar que el tenantId coincide con el tenantId del egreso
+  const egresoExistente = await Egreso.findOne({ _id: egresoId, tenantId });
 
-  if (!egreso) {
-    throw new Error("Egreso no encontrado");
-  } else if (egreso.tenantId !== tenantId) {
-    throw new Error("Tenant incorrecto");
+  if (!egresoExistente) {
+    throw new Error("TenantId proporcionado no coincide con ningun Egreso en la base de datos");
   }
 
-  return await Egreso.findOneAndDelete({ _id: egresoId, tenantId });
+  const egresoEliminado = await Egreso.findOneAndDelete({ _id: egresoId, tenantId });
+  return egresoEliminado;
+  } catch (error) {
+    if (error.name === 'CastError' && error.path === '_id') {
+      throw new Error("_id proporcionado no es válido o no se encontro en la base de datos");
+    } else {
+      throw error; // Propaga el error para que sea manejado en el controlador
+    }
+  }
 };
 
 
 const modificarEgresoPorId = async (egresoId, nuevosDatos, tenantId) => {
+
+  try {
+    // Verificar que el _id del egreso y el tenantId coincidan
+   const egresoExistente = await Egreso.findOne({ _id: egresoId, tenantId });
+
+   if (!egresoExistente) {
+    throw new Error("TenantId proporcionado no existe o no coincide con _id del Egreso a modificar");
+  }
   const egresoModificado =  await Egreso.findOneAndUpdate(
     { _id: egresoId, tenantId },
     nuevosDatos,
     { new: true }
   );
+
+  // Si no se encuentra el egreso, lanzar un error
   if (!egresoModificado) {
     throw new Error("egreso no encontrado");
   }
 
   return egresoModificado;
+
+  } catch (error) {
+    if (error.name === 'CastError' && error.path === '_id') {
+      throw new Error("_id proporcionado no es válido o no se encontro en la base de datos");
+    } else {
+      throw error; // Propaga el error para que sea manejado en el controlador
+    }
+  }  
 };
 
 module.exports = {
